Use the virtualenv, Luke
Python images are now using virtualenvs for their dependencies. Although
it's easy to install dependencies system wide with pip and docker, I was
having problems with DistributionNotFound errors.

The move to google/python is because of a weird bug with current
debian:jessie|wheezy images that cause retrieval of HTTPS urls with
python fail.