Add image for phabricator daemons
Phabricator daemons are not made to be run on the foreground. They are
supposed to be managed by `phd start|stop|restart`, which just spawns
new processes and quits.

There is one supported way so far to run the daemons without
daemonizing them: using `phd debug daemon-name`. The problem with this
is that it is equivalent to starting the daemons with `--trace` and
`--verbose`, which makes for a lot of (most times) useless output.

To avoid cluttering the disk with logs that won't be useful, we resort
to the normal `phd restart` approach, and specify a couple extra flags
to make sure supervisor will keep running even if `phd restart` exists
(which it will).