Delay initialization of phabricator's daemons
This is to give some time for the database to start. If this container
tries to start before the database is running, they'll simply fail,
which is not really exciting. The 5 second delay is small, but should
be enough for most purposes.

What I _should_ be doing though is to test the database connection
before trying to run the daemons. Can connect? start daemons. Can't?
wait a bit and try again.