Make sshd log directly to a file, withoug rsyslogd
Also, generate missing keys using `ssh-keygen -A`. Ideally `/etc/ssh/`
can be a volume where keys are stored so the fingerprints won't change
between restarts, or are even baked into the image. There is no single
clear solution to having the keys available to the image in a secure
fashion out of the box, so just create the keys and leave the user to
figure out his preferred methods.