Do not revalidate php code
opcache.validate_timestamps can be used to have php revalidate the code
on disk every few few seconds. since this runs on a docker image and
restarting the image will be necessary to have a new version of
phabricator running, it can be disabled.

ps: this was a warning from phabricator:

https://secure.phabricator.com/diffusion/P/browse/master/src/applications/cache/spec/PhabricatorOpcodeCacheSpec.php;0449a07f537d1df22ce10cd396d1ba9067c40ac1$153-159